+++ cmd/es-utils/utils/optimism_encoder_test.go
@@ -1,7 +1,7 @@
 // Copyright 2022-2023, EthStorage.
 // For license information, see https://github.com/ethstorage/es-node/blob/main/LICENSE
 
-package encoder
+package utils
 
 import (
 	"bytes"
