+++ ethstorage/node/es_api.go
@@ -10,8 +10,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/log"
-	op "github.com/ethstorage/go-ethstorage/ethstorage/encoder"
-
 	"github.com/ethstorage/go-ethstorage/cmd/es-utils/utils"
 	"github.com/ethstorage/go-ethstorage/ethstorage"
 	"github.com/ethstorage/go-ethstorage/ethstorage/downloader"
@@ -29,7 +27,7 @@ type DecodeType uint64
 const (
 	RawData DecodeType = iota
 	PaddingPer31Bytes
-	OptimismCompactBlob
+	OptimismCompact
 )
 
 func NewESAPI(config *RPCConfig, sm *ethstorage.StorageManager, dl *downloader.Downloader, log log.Logger) *esAPI {
@@ -71,9 +69,9 @@ func (api *esAPI) GetBlob(kvIndex uint64, blobHash common.Hash, decodeType Decod
 
 	if decodeType == PaddingPer31Bytes {
 		ret = utils.DecodeBlob(blob)
-	} else if decodeType == OptimismCompactBlob {
+	} else if decodeType == OptimismCompact {
 		var err error
-		if ret, err = op.ToData(blob); err != nil {
+		if ret, err = utils.ToData(blob); err != nil {
 			return nil, err
 		}
 	}
